move more logic into workerd-integration
