+++ packages/qwik/src/optimizer/src/plugins/vite-server.ts
@@ -13,14 +13,7 @@ import clickToComponent from './click-to-component.html?raw';
 import perfWarning from './perf-warning.html?raw';
 import errorHost from './error-host.html?raw';
 
-import {
-  RequestEvLoaders,
-  RequestEvMode,
-  RequestEvQwikSerializer,
-  RequestEvRoute,
-  RequestEvTrailingSlash,
-} from 'packages/qwik-city/middleware/request-handler/request-event';
-import { getWorkerdHandler } from './workerd-integration';
+import { createWorkerdIncomingMessage, getWorkerdHandler } from './workerd-integration';
 
 function getOrigin(req: IncomingMessage) {
   const { PROTOCOL_HEADER, HOST_HEADER } = process.env;
@@ -129,19 +122,14 @@ export async function configureDevServer(
           });
         });
 
-        const msg = {
-          headers: {},
-        } as IncomingMessage;
-        msg.url = req.url;
-        msg.method = 'GET';
-
-        const srcBase = opts.srcDir
-          ? path.relative(opts.rootDir, opts.srcDir).replace(/\\/g, '/')
-          : 'src';
-
-        const serializedRenderOpts = getSerializedRenderOpts(serverData, isClientDevOnly, manifest);
-        msg.headers['x-workerd-rendering-opts'] = serializedRenderOpts;
-        msg.headers['x-workerd-src-base'] = JSON.stringify(srcBase);
+        const msg = createWorkerdIncomingMessage(
+          req,
+          opts,
+          path,
+          serverData,
+          isClientDevOnly,
+          manifest
+        );
 
         const resp = await workerdHandler(msg);
         // let's just read the whole stream here (this can be improved later)
@@ -394,52 +382,3 @@ export const getSymbolHash = (symbolName: string) => {
   }
   return symbolName;
 };
-
-/**
- * RenderOpts (of type RenderToStreamOptions) is not serializable so we need try to generate a
- * serialized renderOpts (it needs to be serialized so that it can be passed to workerd) in the
- * process we throw away anything that can't be serialized, the result seems to still produce a
- * working app, but we do need to understand the implications here, are the things that we filter
- * out here not needed for rendering the html?
- */
-function getSerializedRenderOpts(
-  serverData: Record<string, any>,
-  isClientDevOnly: boolean,
-  manifest: QwikManifest
-) {
-  const { qwikcity, ...serializableServerData } = serverData;
-  const { ev, ...serializableQwikcity } = qwikcity;
-
-  const renderOpts = {
-    debug: true,
-    locale: serverData.locale,
-    snapshot: !isClientDevOnly,
-    manifest: isClientDevOnly ? undefined : manifest,
-    prefetchStrategy: null,
-    serverData: {
-      ...serializableServerData,
-      qwikcity: serializableQwikcity,
-      ev: {
-        [RequestEvLoaders]: ev[RequestEvLoaders],
-        [RequestEvMode]: ev[RequestEvMode],
-        [RequestEvQwikSerializer]: ev[RequestEvQwikSerializer],
-        [RequestEvRoute]: ev[RequestEvRoute],
-        [RequestEvTrailingSlash]: ev[RequestEvTrailingSlash],
-        basePathname: ev.basePathname,
-        method: ev.method,
-        params: ev.params,
-        pathname: ev.pathname,
-        platform: {
-          ssr: true,
-          node: ev.platform.node,
-        },
-        url: ev.url,
-      },
-    },
-    containerAttributes: {
-      ...serverData.containerAttributes,
-    },
-  };
-  const serializedRenderOpts = JSON.stringify(renderOpts);
-  return serializedRenderOpts;
-}
